avanzando alertas

+++ database/migrations/2023_03_25_213007_create_games_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->integer("anyoLanzamiento");
             $table->json("generos");
             $table->json("plataformas");
-            $table->integer("precio");
+            $table->double("precio");
             $table->binary("imagen")->nullable();
             $table->timestamps();
         });
